feat(tag): add batch gpt tagging script for charts
