+++ adminSiteServer/apiRouter.ts
@@ -14,12 +14,7 @@ import {
     DATA_API_URL,
 } from "../settings/serverSettings.js"
 import { expectInt, isValidSlug } from "../serverUtils/serverUtil.js"
-import {
-    OldChart,
-    Chart,
-    getGrapherById,
-    PUBLIC_TAG_PARENT_IDS,
-} from "../db/model/Chart.js"
+import { OldChart, Chart, getGrapherById } from "../db/model/Chart.js"
 import { Request, Response, CurrentUser } from "./authentication.js"
 import {
     getMergedGrapherConfigForVariable,
@@ -52,6 +47,7 @@ import {
     OwidChartDimensionInterface,
     DimensionProperty,
     Tag,
+    TaggableType,
 } from "@ourworldindata/utils"
 import {
     GrapherInterface,
@@ -97,8 +93,6 @@ import { Link } from "../db/model/Link.js"
 import { In } from "typeorm"
 import { GIT_CMS_DIR } from "../gitCms/GitCmsConstants.js"
 import { logErrorAndMaybeSendToBugsnag } from "../serverUtils/errorLog.js"
-import { TaggableType } from "../adminSiteClient/EditableTags.js"
-import { OpenAI } from "openai"
 
 const apiRouter = new FunctionalRouter()
 const explorerAdminServer = new ExplorerAdminServer(GIT_CMS_DIR)
@@ -2610,66 +2604,19 @@ apiRouter.post(
 apiRouter.get(
     `/gpt/suggest-topics/${TaggableType.Charts}/:chartId.json`,
     async (req: Request, res: Response): Promise<Record<"topics", Tag[]>> => {
-        const openai = new OpenAI()
-        const chart = await Chart.findOneBy({
-            id: parseIntOrUndefined(req.params.chartId),
-        })
-        if (!chart)
+        const chartId = parseIntOrUndefined(req.params.chartId)
+        if (!chartId) throw new JsonError(`Invalid chart ID`, 400)
+
+        const topics = await Chart.getGptTopicSuggestions(chartId)
+
+        if (!topics.length)
             throw new JsonError(
-                `No chart found for id ${req.params.chartId}`,
+                `No GPT topic suggestions found for chart ${chartId}`,
                 404
             )
 
-        const topics: Tag[] = await db.queryMysql(`
-            SELECT t.id, t.name
-            FROM tags t 
-            WHERE t.isTopic IS TRUE
-            AND t.parentId IN (${PUBLIC_TAG_PARENT_IDS.join(",")})
-        `)
-
-        if (!topics.length) throw new JsonError("No topics found", 404)
-
-        const prompt = `
-            You will be provided with the chart metadata (delimited with XML tags),
-            as well as a list of possible topics (delimited with XML tags).
-            Classify the chart into two of the provided topics.
-            <chart>
-                <title>${chart.config.title}</title>
-                <description>${chart.config.subtitle}</description>
-                <listed-on>${chart.config.originUrl}</listed-on>
-            </chart>
-            <topics>
-                ${topics.map(
-                    (topic) => `<topic id=${topic.id}>${topic.name}</topic>\n`
-                )}
-            </topics>
-            
-            Respond with the two categories you think best describe the chart. 
-            
-            Format your response as follows:
-            [
-                { "id": 1, "name": "Topic 1" },
-                { "id": 2, "name": "Topic 2" }
-            ]`
-
-        const completion = await openai.chat.completions.create({
-            messages: [{ role: "user", content: prompt }],
-            model: "gpt-3.5-turbo",
-        })
-
-        const json = completion.choices[0]?.message?.content
-        if (!json) throw new JsonError("No response from GPT", 500)
-
-        const selectedTopics: Tag[] = JSON.parse(json)
-
-        // We only want to return topics that are in the list of possible
-        // topics, in case of hallucinations
-        const confirmedTopics = selectedTopics.filter((topic) =>
-            topics.map((t) => t.id).includes(topic.id)
-        )
-
         return {
-            topics: confirmedTopics,
+            topics,
         }
     }
 )
