+++ package.json
@@ -6,6 +6,7 @@
     },
     "packageManager": "yarn@3.3.0",
     "scripts": {
+        "batchTagWithGpt": "node --enable-source-maps ./itsJustJavascript/baker/batchTagWithGpt.js",
         "buildAndDeploySite": "node --enable-source-maps ./itsJustJavascript/baker/buildAndDeploySite.js",
         "buildCoverage": "jest --coverage=true --coverageProvider=v8",
         "buildLocalBake": "node --enable-source-maps ./itsJustJavascript/baker/buildLocalBake.js",
